Changed max admissible linear velocity of drone. Changed decimal variables from float to double for better precision. Created arena_offline that moves the drone per callback call: a kind of sequential interaction. In other words, each physics update is called per callback call. Added my_controller3.cpp that exemplifies a sequential publishing of commands.

+++ include/geometry.h
@@ -6,20 +6,20 @@
 #include <vector>
 #include <string>
 
-float distance_between_points(float, float, float, float);
+double distance_between_points(double, double, double, double);
 
 class Point2d{
 	private:
 		Eigen::Vector2f data;
 		Eigen::Rotation2D<float> rotation;
 	public:
-		Point2d(float x, float y);
-		float get_x();
-		void set_x(float x);
-		float get_y();
-		void set_y(float y);
+		Point2d(double x, double y);
+		double get_x();
+		void set_x(double x);
+		double get_y();
+		void set_y(double y);
 		std::string tostring();
-		void rotate(float degrees);
+		void rotate(double degrees);
 }; 
 
 class Circle;
@@ -28,71 +28,71 @@ class Line{
 	protected:
 		double denominator, numerator, slope, intercept, x1, y1, x2, y2;
 		bool horizontal, vertical;
-		float ordered_points_x_[2], ordered_points_y_[2];
+		double ordered_points_x_[2], ordered_points_y_[2];
 	public:
-		Line(float x1,float y1,float x2, float y2);
+		Line(double x1,double y1,double x2, double y2);
 		std::string tostring();
-		float get_slope();
-		float get_intercept();
-		void set_x1(float x1);
-		void set_x2(float x2);
-		void set_y1(float y1);
-		void set_y2(float y2);
-		float get_x1();
-		float get_x2();
-		float get_y1();
-		float get_y2();
-		float* get_ordered_points_x();
-		float* get_ordered_points_y();
-		void set_points(float, float, float, float);
-		bool intersects(float x,float y);
+		double get_slope();
+		double get_intercept();
+		void set_x1(double x1);
+		void set_x2(double x2);
+		void set_y1(double y1);
+		void set_y2(double y2);
+		double get_x1();
+		double get_x2();
+		double get_y1();
+		double get_y2();
+		double* get_ordered_points_x();
+		double* get_ordered_points_y();
+		void set_points(double, double, double, double);
+		bool intersects(double x,double y);
 		bool is_vertical();
 		bool is_horizontal();
-		std::tuple<float, float> get_midpoint();
-		float get_length();
-		virtual std::tuple<bool, float, float> intersects_line(Line other);
-		virtual std::tuple<bool, float, float, float, float> intersects_circle(Circle circle);
+		std::tuple<double, double> get_midpoint();
+		double get_length();
+		virtual std::tuple<bool, double, double> intersects_line(Line other);
+		virtual std::tuple<bool, double, double, double, double> intersects_circle(Circle circle);
 };
 
 class LineSegment : public Line{
 	public:
-		LineSegment(float x1, float y1, float x2, float y2);
-		std::tuple<bool, float, float, float, float> intersects_circle(Circle circle) override final;
-		std::tuple<bool, float, float> intersects_line(Line other) override final;
+		LineSegment(double x1, double y1, double x2, double y2);
+		std::tuple<bool, double, double, double, double> intersects_circle(Circle circle) override final;
+		std::tuple<bool, double, double> intersects_line(Line other) override final;
 };
 
 class Square{
 	protected:
-		float length, xc, yc, angle;
+		double length, xc, yc, angle;
 		std::vector<Point2d> corners;
 		std::vector<LineSegment> edges;	
 	public:
 		std::vector<LineSegment> get_edges();
-		Square(float length, float xc, float yc, float angle);
+		Square(double length, double xc, double yc, double angle);
 		std::string tostring();
-		float get_xc();
-		float get_yc();
-		float get_angle();
-		float get_length();
+		double get_xc();
+		double get_yc();
+		double get_angle();
+		double get_length();
 };
 
 class Circle{
 	protected:
-		float radius, diameter, radius_sqr, xc, yc;	
+		double radius, diameter, radius_sqr, xc, yc;	
 	public:
-		Circle(float xc, float yc, float radius);
+		Circle(double xc, double yc, double radius);
 		std::string tostring();
-		float get_xc();
-		float get_yc();
-		void set_xc(float x); 
-		void set_yc(float y);
-		float get_radius();
-		float get_diameter();
-		float equation(float x, float y);
-		bool intersects(float x, float y);
-		bool inside(float x, float y);
-		bool outside(float x, float y);
+		double get_xc();
+		double get_yc();
+		void set_xc(double x); 
+		void set_yc(double y);
+		double get_radius();
+		double get_diameter();
+		double equation(double x, double y);
+		bool intersects(double x, double y);
+		bool inside(double x, double y);
+		bool outside(double x, double y);
 		bool intersects_circle(Circle other);
-		std::tuple<bool, float, float, float, float> intersects_line(Line line);
+		std::tuple<bool, double, double, double, double> intersects_line(Line line);
 };
 #endif // GEOMETRY_H
