+++ src/geometry.cpp
@@ -8,23 +8,23 @@
 #define INTERSECTION_POINT_TOLERANCE 0.01
 
 //Point2d
-Point2d::Point2d(float x, float y){
+Point2d::Point2d(double x, double y){
 	this->data << x,y;
 }
 
-float Point2d::get_x(){return this->data[0];}
-void Point2d::set_x(float x){this->data[0] = x;}
-float Point2d::get_y(){return this->data[1];}
-void Point2d::set_y(float y){this->data[1] = y;}
+double Point2d::get_x(){return this->data[0];}
+void Point2d::set_x(double x){this->data[0] = x;}
+double Point2d::get_y(){return this->data[1];}
+void Point2d::set_y(double y){this->data[1] = y;}
 std::string Point2d::tostring(){return "Point (" + std::to_string(data[0]) + "," + std::to_string(data[1])+")";}
-void Point2d::rotate(float degrees){
+void Point2d::rotate(double degrees){
 	this->rotation.angle() = degrees * TO_RAD;
 	this->data = this->rotation.toRotationMatrix() * this->data;
 	return;
 }
 
 // Line	
-Line::Line(float x1,float y1,float x2, float y2){
+Line::Line(double x1,double y1,double x2, double y2){
 	this->horizontal = false;
 	this->vertical = false;
 	this->set_points(x1,y1,x2,y2);
@@ -42,34 +42,34 @@ std::string Line::tostring(){
 	+"\n";
 }
 
-float Line::get_slope(){return this->slope;}
+double Line::get_slope(){return this->slope;}
 
-float Line::get_intercept(){return this->intercept;}
+double Line::get_intercept(){return this->intercept;}
 
-void Line::set_x1(float x1){this->x1 = x1;}
+void Line::set_x1(double x1){this->x1 = x1;}
 
-void Line::set_x2(float x2){this->x2 = x2;}
+void Line::set_x2(double x2){this->x2 = x2;}
 
-void Line::set_y1(float y1){this->y1 = y1;}
+void Line::set_y1(double y1){this->y1 = y1;}
 
-void Line::set_y2(float y2){this->y2 = y2;}
+void Line::set_y2(double y2){this->y2 = y2;}
 
-float Line::get_x1(){return this->x1;}
-float Line::get_y1(){return this->y1;}
-float Line::get_x2(){return this->x2;}
-float Line::get_y2(){return this->y2;}
+double Line::get_x1(){return this->x1;}
+double Line::get_y1(){return this->y1;}
+double Line::get_x2(){return this->x2;}
+double Line::get_y2(){return this->y2;}
 
-float* Line::get_ordered_points_x()
+double* Line::get_ordered_points_x()
 {
 	return ordered_points_x_;
 }
 
-float* Line::get_ordered_points_y()
+double* Line::get_ordered_points_y()
 {
 	return ordered_points_y_;
 }
 		
-void Line::set_points(float x1, float y1, float x2, float y2){
+void Line::set_points(double x1, double y1, double x2, double y2){
 	this->x1 = x1;
 	this->y1 = y1;
 	this->x2 = x2;
@@ -82,8 +82,8 @@ void Line::set_points(float x1, float y1, float x2, float y2){
 	ordered_points_x_[1] = x2;
 	ordered_points_y_[0] = y1;
 	ordered_points_y_[1] = y2;
-	std::sort(ordered_points_x_, ordered_points_x_ + individual_size<float*>(ordered_points_x_));
-	std::sort(ordered_points_y_, ordered_points_y_ + individual_size<float*>(ordered_points_y_));
+	std::sort(ordered_points_x_, ordered_points_x_ + individual_size<double*>(ordered_points_x_));
+	std::sort(ordered_points_y_, ordered_points_y_ + individual_size<double*>(ordered_points_y_));
 	// If line is not vertical
 	if (this->denominator != 0 || abs(this->denominator) >= DENOMINATOR_TOLERANCE){
 		this->slope = numerator / (double)denominator;
@@ -105,23 +105,23 @@ void Line::set_points(float x1, float y1, float x2, float y2){
 	}
 }
 
-bool Line::intersects(float x,float y){return y -  this->slope * x - this->intercept <= INTERSECTION_POINT_TOLERANCE;}
+bool Line::intersects(double x,double y){return y -  this->slope * x - this->intercept <= INTERSECTION_POINT_TOLERANCE;}
 
 bool Line::is_vertical(){return this->vertical;}
 
 bool Line::is_horizontal(){return this->horizontal;}
 
-std::tuple<float, float> Line::get_midpoint(){
+std::tuple<double, double> Line::get_midpoint(){
 	return {
 		0.5 * (this->x1 + this->x2),
 		0.5 * (this->y1 + this->y2)
 	};
 }
 
-float Line::get_length(){
+double Line::get_length(){
 	return sqrt(pow(x2 - x1, 2) + pow(y2 - y1, 2));
 }
-std::tuple<bool, float, float> Line::intersects_line(Line other){
+std::tuple<bool, double, double> Line::intersects_line(Line other){
 	if (this->slope == other.get_slope()) return {false,0,0};
 	else if(!this->vertical && other.vertical) return {true, other.get_intercept(), this->slope * other.get_intercept() + this->intercept};
 	else if(this->vertical && !other.vertical) return {true, this->intercept, other.get_slope() * this->intercept + other.get_intercept()};
@@ -135,32 +135,32 @@ std::tuple<bool, float, float> Line::intersects_line(Line other){
 
 }
 
-std::tuple<bool, float, float, float, float> Line::intersects_circle(Circle circle){ //x1, y1, x2, y2
-	float discriminant;
+std::tuple<bool, double, double, double, double> Line::intersects_circle(Circle circle){ //x1, y1, x2, y2
+	double discriminant;
 	if (this->vertical){
 		discriminant = pow(circle.get_radius(),2) - pow(this->intercept - circle.get_xc(),2);
 		if (discriminant <= 0) return {false, 0,0,0,0};
-		float sqrt_d = sqrt(discriminant);
+		double sqrt_d = sqrt(discriminant);
 		return {true, this->intercept, circle.get_yc() + sqrt_d, this-> intercept, circle.get_yc() - sqrt_d};
 	}
 	else{
-		float slope_sqr = pow(this->slope,2);
+		double slope_sqr = pow(this->slope,2);
 		discriminant = pow(circle.get_radius(), 2) * (1 + slope_sqr) - pow(circle.get_yc() - this->slope * circle.get_xc() - this->intercept,2);
 		if(discriminant <= 0) return {false, 0,0,0,0};
-		float den = 1 + slope_sqr;
-		float sqrt_d = sqrt(discriminant);
-		float a = circle.get_xc() + circle.get_yc() * this->slope - this->intercept * this->slope;
-		float b = this->intercept + circle.get_xc() * this->slope + circle.get_yc() * slope_sqr;
+		double den = 1 + slope_sqr;
+		double sqrt_d = sqrt(discriminant);
+		double a = circle.get_xc() + circle.get_yc() * this->slope - this->intercept * this->slope;
+		double b = this->intercept + circle.get_xc() * this->slope + circle.get_yc() * slope_sqr;
 		return {true, (a + sqrt_d) / den, (b + this->slope * sqrt_d) / den, (a - sqrt_d) / den, (b - this->slope * sqrt_d) / den};
 	}
 }
 
 //Line Segment
-LineSegment::LineSegment(float x1, float y1, float x2, float y2) : Line(x1,x2,y1,y2){}
+LineSegment::LineSegment(double x1, double y1, double x2, double y2) : Line(x1,x2,y1,y2){}
 
-std::tuple<bool, float, float, float, float> LineSegment::intersects_circle(Circle circle)
+std::tuple<bool, double, double, double, double> LineSegment::intersects_circle(Circle circle)
 {
-	std::tuple<bool, float, float, float, float> points = Line::intersects_circle(circle);
+	std::tuple<bool, double, double, double, double> points = Line::intersects_circle(circle);
 	if(std::get<0>(points))
 	{
 		if(!in_between(ordered_points_x_[0], std::get<1>(points), ordered_points_x_[1])
@@ -173,9 +173,9 @@ std::tuple<bool, float, float, float, float> LineSegment::intersects_circle(Circ
 	}
 	return points;
 }
-std::tuple<bool, float, float> LineSegment::intersects_line(Line other)
+std::tuple<bool, double, double> LineSegment::intersects_line(Line other)
 {
-	std::tuple<bool, float, float> points = Line::intersects_line(other);
+	std::tuple<bool, double, double> points = Line::intersects_line(other);
 	if(std::get<0>(points))
 	{
 		if(!in_between(ordered_points_x_[0], std::get<1>(points), ordered_points_x_[1]) 
@@ -188,60 +188,60 @@ std::tuple<bool, float, float> LineSegment::intersects_line(Line other)
 }
 
 //Circle
-Circle::Circle(float xc, float yc, float radius) : xc(xc), yc(yc), radius(radius), radius_sqr(pow(this->radius,2)) {
+Circle::Circle(double xc, double yc, double radius) : xc(xc), yc(yc), radius(radius), radius_sqr(pow(this->radius,2)) {
 	this->diameter = radius * 2.0;
 }
 
 std::string Circle::tostring(){return "Circle with xc: " + std::to_string(this->xc) + ", yc: " + std::to_string(this->yc)+", radius: " + std::to_string(this->radius);}
 
-float Circle::get_xc(){return this->xc;}
+double Circle::get_xc(){return this->xc;}
 
-float Circle::get_yc(){return this->yc;}
+double Circle::get_yc(){return this->yc;}
 
-void Circle::set_xc(float x){this->xc = x;}
+void Circle::set_xc(double x){this->xc = x;}
 
-void Circle::set_yc(float y){this->yc = y;}
+void Circle::set_yc(double y){this->yc = y;}
 
-float Circle::get_radius(){return this->radius;}
+double Circle::get_radius(){return this->radius;}
 
-float Circle::get_diameter(){return this->diameter;}
+double Circle::get_diameter(){return this->diameter;}
 
-float Circle::equation(float x, float y){return pow(x - this->xc, 2) + pow(y - this->yc, 2);}
+double Circle::equation(double x, double y){return pow(x - this->xc, 2) + pow(y - this->yc, 2);}
 
-bool Circle::intersects(float x, float y){return this->radius_sqr - INTERSECTION_POINT_TOLERANCE <= equation(x,y) < this->radius_sqr - INTERSECTION_POINT_TOLERANCE;}
+bool Circle::intersects(double x, double y){return this->radius_sqr - INTERSECTION_POINT_TOLERANCE <= equation(x,y) < this->radius_sqr - INTERSECTION_POINT_TOLERANCE;}
 
-bool Circle::inside(float x, float y){return equation(x,y) <= this->radius_sqr;}
+bool Circle::inside(double x, double y){return equation(x,y) <= this->radius_sqr;}
 
-bool Circle::outside(float x, float y){return !inside(x,y);}
+bool Circle::outside(double x, double y){return !inside(x,y);}
 
 bool Circle::intersects_circle(Circle other){
-	float d = sqrt(equation(other.get_xc(), other.get_yc()));
+	double d = sqrt(equation(other.get_xc(), other.get_yc()));
 	return (this->radius + other.get_radius()) > d && (d > abs(this->radius - other.get_radius()));
 }
 
 
-std::tuple<bool, float, float, float, float> Circle::intersects_line(Line line){
-	float discriminant;
+std::tuple<bool, double, double, double, double> Circle::intersects_line(Line line){
+	double discriminant;
 	if (line.is_vertical()){
 		discriminant = pow(this->radius,2) - pow(line.get_intercept() - this->xc, 2);
 		if (discriminant <= 0) return {false, 0,0,0,0};
-		float sqrt_d = sqrt(discriminant);
+		double sqrt_d = sqrt(discriminant);
 		return {true, line.get_intercept(), this->yc + sqrt_d, line.get_intercept(), this->yc - sqrt_d};
 	}
 	else{
-		float slope_sqr = pow(line.get_slope(),2);
+		double slope_sqr = pow(line.get_slope(),2);
 		discriminant = pow(this->radius, 2) * (1 + slope_sqr) - pow(this->yc - line.get_slope() * this->xc - line.get_intercept(), 2);
 		if(discriminant <= 0) return {false,0,0,0,0};
-		float den = 1 + slope_sqr;
-		float sqrt_d = sqrt(discriminant);
-		float a = this->xc + this->yc * line.get_slope() - line.get_intercept() * line.get_slope();
-		float b = line.get_intercept() + this->xc * line.get_slope() + this->yc * slope_sqr;
+		double den = 1 + slope_sqr;
+		double sqrt_d = sqrt(discriminant);
+		double a = this->xc + this->yc * line.get_slope() - line.get_intercept() * line.get_slope();
+		double b = line.get_intercept() + this->xc * line.get_slope() + this->yc * slope_sqr;
 		return {true, (a + sqrt_d) / den, (b + line.get_slope() * sqrt_d) / den, (a - sqrt_d) / den, (b - line.get_slope() * sqrt_d) / den};
 	}
 }
 
 // Square
-Square::Square(float length, float xc, float yc, float angle) : length(length), xc(xc), yc(yc), angle(angle){
+Square::Square(double length, double xc, double yc, double angle) : length(length), xc(xc), yc(yc), angle(angle){
 	// Corners
 	this->corners.push_back(Point2d(xc - length * 0.5, yc + length * 0.5));
 	this->corners.push_back(Point2d(xc + length * 0.5, yc + length * 0.5));
@@ -270,21 +270,21 @@ std::vector<LineSegment> Square::get_edges(){
 	return this->edges;
 }
 
-float distance_between_points(float x1, float y1, float x2, float y2){
+double distance_between_points(double x1, double y1, double x2, double y2){
 	return sqrt(pow(x2-x1,2)+pow(y2-y1,2));
 }
 
-float Square::get_xc(){
+double Square::get_xc(){
 	return this->xc;
 }
-float Square::get_yc(){
+double Square::get_yc(){
 	return this->yc;
 }
 
-float Square::get_angle(){
+double Square::get_angle(){
 	return this->angle;
 }
 
-float Square::get_length(){
+double Square::get_length(){
 	return this->length;
 }
